backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "openai>=2.14.0",
    "pdfplumber>=0.11.8",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "qdrant-client>=1.16.2",
    "uvicorn[standard]>=0.40.0",
]
